Module ink_storage_derive::spread_layout

Functions

footprint
Generates the tokens for the SpreadLayout footprint of some type.

max_n
Generates the tokens to compute the maximum of the numbers given via their token streams at compilation time.

requires_deep_clean_up
Generates the tokens for the SpreadLayout REQUIRES_DEEP_CLEAN_UP constant for the given structure.

spread_layout_derive
Derives ink_storage's SpreadLayout trait for the given struct or enum.

spread_layout_enum_derive
SpreadLayout derive implementation for enum types.

spread_layout_struct_derive
SpreadLayout derive implementation for struct types.